+++ src/test/script_p2sh_tests.cpp
@@ -11,6 +11,7 @@
 #include <script/sign.h>
 #include <script/signingprovider.h>
 #include <test/util/setup_common.h>
+#include <test/util/transaction_utils.h>
 #include <validation.h>
 
 #include <vector>
